# packages/daemon/openapi.yaml
openapi: 3.0.3
info:
  title: orch8 API
  description: |
    REST + WebSocket API for the orch8 orchestration platform.

    ## Authentication

    Two authentication paths are supported on all endpoints (except `GET /api/health`):

    **Agent Authentication** — Provide both `X-Agent-Id` and `X-Project-Id` headers. The agent is validated against the database.

    **Admin Authentication** — Requests from localhost (`127.0.0.1`, `::1`) without agent headers are treated as admin. Admins bypass all permission checks.
  version: 0.0.0
  license:
    name: Private
servers:
  - url: http://localhost:3847
    description: Local daemon
...
components:
  securitySchemes:
    AgentAuth:
      type: apiKey
      in: header
      name: X-Agent-Id
      description: Agent ID header (must be paired with X-Project-Id)
    ProjectScope:
      type: apiKey
      in: header
      name: X-Project-Id
      description: Project ID scope header
    RunContext:
      type: apiKey
      in: header
      name: X-Run-Id
      description: Optional run ID for correlation
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: not_found
        message:
          type: string
          example: Task not found
    ValidationError:
      type: object
      required: [error, details]
      properties:
        error:
          type: string
          enum: [validation_error]
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              path:
                type: array
                items:
                  type: string
              message:
                type: string
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true

    # --- Project ---
...
    Task:
      type: object
      properties:
        id:
          type: string
          example: task_abc123
        title:
          type: string
        description:
          type: string
        projectId:
          type: string
        column:
          $ref: "#/components/schemas/TaskColumn"
        taskType:
          $ref: "#/components/schemas/TaskType"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assignee:
          type: string
          nullable: true
        brainstormStatus:
          $ref: "#/components/schemas/BrainstormStatus"
        branch:
          type: string
          nullable: true
        worktreePath:
          type: string
          nullable: true
        verificationResult:
          $ref: "#/components/schemas/VerificationResult"
        verifierReport:
          type: string
          nullable: true
        refereeVerdict:
          type: string
          nullable: true
        executionRunId:
          type: string
          nullable: true
        executionAgentId:
          type: string
          nullable: true
        retryCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateQuickTask:
      type: object
      required: [title, projectId, taskType]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          maxLength: 10000
          default: ""
        projectId:
          type: string
          minLength: 1
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assignee:
          type: string
        taskType:
          type: string
          enum: [quick]
        dependsOn:
          type: array
          items:
            type: string
          description: Task IDs this task depends on; created in a blocked state until all deps complete
    CreateBrainstormTask:
      type: object
      required: [title, projectId, taskType]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          maxLength: 10000
          default: ""
        projectId:
          type: string
          minLength: 1
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assignee:
          type: string
        taskType:
          type: string
          enum: [brainstorm]
        dependsOn:
          type: array
          items:
            type: string
          description: Task IDs this task depends on; created in a blocked state until all deps complete
    CreateTask:
      oneOf:
        - $ref: "#/components/schemas/CreateQuickTask"
        - $ref: "#/components/schemas/CreateBrainstormTask"
      discriminator:
        propertyName: taskType
        mapping:
          quick: "#/components/schemas/CreateQuickTask"
          brainstorm: "#/components/schemas/CreateBrainstormTask"
    UpdateTask:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          maxLength: 10000
        column:
          $ref: "#/components/schemas/TaskColumn"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assignee:
          type: string
          nullable: true
        autoCommit:
          type: boolean
        autoPr:
          type: boolean
        branch:
          type: string
          nullable: true
        worktreePath:
          type: string
          nullable: true
        mcpTools:
          type: array
          items:
            type: string

    # --- Agent ---
...
    TaskType:
      type: string
      enum: [quick, brainstorm]
    TaskColumn:
      type: string
      enum: [backlog, blocked, in_progress, review, verification, done]
    TaskPriority:
      type: string
      enum: [high, medium, low]
      default: medium
    BrainstormStatus:
      type: string
      enum: [active, idle, ready, expired]
      nullable: true
...
    VerificationResult:
      type: string
      enum: [pass, fail, partial]
      nullable: true
...
paths:
...
  /api/tasks:
    post:
      tags: [Tasks]
      summary: Create task
      description: |
        Creates a quick or brainstorm task. The request body is a discriminated union on `taskType`.
        Requires `create_task` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTask"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: No create_task permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
        - name: column
          in: query
          schema:
            $ref: "#/components/schemas/TaskColumn"
        - name: taskType
          in: query
          schema:
            $ref: "#/components/schemas/TaskType"
        - name: assignee
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /api/tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [Tasks]
      summary: Update task
      description: |
        If `column` is present, the request is redirected to the lifecycle transition service.
        Requires `assign_task` permission.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTask"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error or invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: No assign_task permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Tasks]
      summary: Delete task
      description: |
        Deletes a task. Dependencies and comments are cascaded; run
        references become null. Requires `create_task` permission.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: No create_task permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Task is actively executing or part of a pipeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/{id}/transition:
    post:
      tags: [Tasks]
      summary: Explicit lifecycle transition
      description: Requires `move_task` permission.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [column]
              properties:
                column:
                  $ref: "#/components/schemas/TaskColumn"
                agentId:
                  type: string
                  description: Defaults to caller's agent ID
                runId:
                  type: string
                  description: Defaults to caller's run ID
      responses:
        "200":
          description: Transitioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: No move_task permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/{id}/complete:
    post:
      tags: [Tasks]
      summary: Signal task completion
      description: |
        For quick tasks, no body needed — transitions to review.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                output:
                  type: string
                  minLength: 1
                  description: Optional completion output
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error or invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/{id}/dependencies:
    post:
      tags: [Tasks]
      summary: Add dependency
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dependsOnId]
              properties:
                dependsOnId:
                  type: string
      responses:
        "201":
          description: Dependency added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Missing dependsOnId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Cycle detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/{id}/dependencies/{depId}:
    delete:
      tags: [Tasks]
      summary: Remove dependency
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: depId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dependency removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /api/tasks/{id}/convert:
    post:
      tags: [Tasks]
      summary: Convert brainstorm to work task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taskType]
              properties:
                taskType:
                  type: string
                  enum: [quick]
      responses:
        "200":
          description: Converted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error or not a brainstorm task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/{id}/checkout:
    post:
      tags: [Tasks]
      summary: Atomic task claim
      description: |
        Agent-driven atomic checkout. Claims the task for the calling agent's
        current run. Agent authentication is required.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Cannot checkout in current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Agent authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Checkout conflict (already held by another run)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tasks/{id}/release:
    post:
      tags: [Tasks]
      summary: Release execution lock
      description: |
        Releases the task's execution lock without completing it. Only the
        agent currently holding the lock may release.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Agent authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller does not hold this task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ============================================================
  # Comments
  # ============================================================
...